Fix ooo.get({ id, meta: true })
Previously this function wasn't checking the `meta` property, which
ssb-db uses to determine whether it should output `data` or `data.value`
without the `key` property. This was causing very strange behavior in
clients that depend on `get()` because this module hooks into ssb-db's
`get()` method, so it's important that ssb-ooo provides the same behavior.